Decoupled BuildContext and functions
