+++ bpy_addon_build/build_context/__init__.py
@@ -1,0 +1,47 @@
+from pathlib import Path
+from typing import List
+from attrs import define
+from bpy_addon_build.api import Api
+from bpy_addon_build.args import Args
+
+from bpy_addon_build.config import Config
+from rich.console import Console
+
+INSTALL_PATHS: List[str] = [
+    "~/AppData/Roaming/Blender Foundation/Blender/{0}/scripts/addons",
+    "~/Library/Application Support/Blender/{0}/scripts/addons",
+    "~/.config/blender/{0}/scripts/addons",
+]
+
+# Must be ignored because Mypy likes
+# to complain about this for some reason
+WORKING_DIR = Path.cwd()  # type: ignore
+
+console = Console()
+
+
+# Must be ignored to pass Mypy as this has
+# an expression of Any, likely due to how
+# attrs works
+@define  # type: ignore
+class BuildContext:
+    """
+    Context of the build environment, from settings to
+    actions to paths, etc.
+
+    Attributes
+    ----------
+    config_path: Path
+        Path to config file
+
+    config: Config
+        Configuration defined by the user
+
+    cli: Args
+        Arguments passed by the user
+    """
+
+    config_path: Path
+    config: Config
+    cli: Args
+    api: Api
