+++ bpy_addon_build/build_context/hook_definitions.py
